bug fixes and handle file paths

+++ connectDB.sh
@@ -3,60 +3,57 @@
 export pat="^[a-zA-Z]+[a-zA-Z0-9[:space:]]*"
 
 function showTables {
-            ls | awk 'BEGIN{FS=" "} { print "- "$1 }';
-         }
+    res=$(ls | awk 'BEGIN{FS=" "} { print "- "$1 }')
+    if [[ $res == "" ]]; then
+        echo "No Tables found"
+    else
+        echo $res
+    fi
+    showMenu
+}
 
 function dropTable {
-  echo -e "Enter Table Name: \c"
-  read name
-  rm $name 
-  rm .$name 
-  if [[ $? == 0 ]]
-  then
-    echo "Table Dropped Successfully"
-  else
-    echo "Error Dropping Table $name"
-  fi
+    read -r -p "Enter Table Name: " name
+    if [ -f "$name" ]; then
+        read -r -p "You are going to drop table $name? Are you sure? [Y/N]: " check
+        if [[ $check = "Y" || $check = "y" ]]; then
+            rm $name
+            rm .$name
+            echo "Table Dropped Successfully"
+        fi
+    else
+        echo "$name table is not existed!"
+    fi
+    showMenu
+
 }
 
 function showMenu() {
     select choice in createTable listTable dropTable insertTable selectTable removeFromTable updateTable "Go back to main menu"; do
         case $choice in
         createTable)
-            . createTable.sh
+            . ../../createTable.sh
             ;;
         listTable)
+            pwd
             echo "listTable"
             showTables
-            select choice in Goback; do
-                case $choice in
-                    Goback)
-                    showMenu
-                    ;;
-                esac
-            done
+            showMenu
             ;;
         dropTable)
             echo "dropTable"
             dropTable
-            select choice in Goback; do
-                case $choice in
-                    Goback)
-                    showMenu
-                    ;;
-                esac
-            done
             ;;
         insertTable)
-            . insertTable.sh
+            . ../../insertTable.sh
             ;;
         selectTable)
             echo "selectTable"
-            . selectTable.sh
+            . ../../selectTable.sh
             ;;
         removeFromTable)
             echo "removeFromTable"
-            . deleteDataFromTable.sh
+            . ../../deleteDataFromTable.sh
             ;;
         updateTable)
             echo "updateTable"
